requirements.txt
pandas 
numpy 
scikit-learn 
xgboost 
ccxt 
requests 
newsapi-python 
google-generativeai 
vectorbt 
pyarrow 
jupyter 
matplotlib 
